+++ src/app/globals.css
@@ -6,3 +6,11 @@ svg {
   width: 20px;
   height: 20px;
 }
+
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
