+++ src/app/page.module.css
@@ -1,12 +1,3 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  /* border: 1px solid; */
-  width: 100%;
-}
-
 .line {
   float: right;
   display: block;
@@ -30,112 +21,4 @@
   color: gray;
   font-weight: 600;
   padding: 10px 0;
-}
-
-/* search */
-
-.searchCover {
-  height: 300px;
-  background: url("/images/cover.webp") no-repeat center;
-  background-size: cover;
-  filter: grayscale(50%);
-
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.search {
-  /* border: 1px solid blue; */
-  /* height: 200px; */
-  /* 바꾸기 */
-  width: 1024px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  justify-content: center;
-  color: #fff;
-  padding: 20px;
-}
-
-.title {
-  font-size: 40px;
-  font-weight: bold;
-}
-
-.input {
-  height: 54px;
-  width: 70%;
-  border-radius: 10px;
-  border: 0;
-  padding-left: 20px;
-  color: black;
-}
-
-.searchIcon {
-  color: gray;
-  width: 23px;
-  height: 23px;
-  position: relative;
-  left: -35px;
-  top: 5px;
-  stroke-width: 2.5px;
-}
-
-.photoList {
-  margin-top: 50px;
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-template-rows: 1fr;
-  gap: 20px;
-}
-@media (max-width: 860px) {
-  .photoList {
-    grid-template-columns: 1fr 1fr 1fr;
-  }
-}
-
-@media (max-width: 650px) {
-  .photoList {
-    grid-template-columns: 1fr 1fr;
-  }
-}
-@media (max-width: 422px) {
-  .photoList {
-    grid-template-columns: 1fr;
-  }
-}
-
-.photoItem {
-  /* border: 1px solid black; */
-  width: 200px;
-  height: 200px;
-  position: relative;
-}
-
-.photoItem > img {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-
-.likeButton {
-  position: absolute;
-  bottom: 5px;
-  right: 5px;
-  width: 25px;
-  height: 25px;
-  stroke-width: 2px;
-  stroke: #fff;
-  fill: red;
-}
-
-.pagingWrapper {
-  border: 1px solid;
-  height: 50px;
-  margin-top: 50px;
-  margin-bottom: 100px;
 }
