+++ src/app/ui/main/photo-item.module.css
@@ -1,0 +1,25 @@
+.photoItem {
+  width: 200px;
+  height: 200px;
+  position: relative;
+  cursor: pointer;
+}
+
+.photoItem > img {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+}
+
+.likeButton {
+  position: absolute;
+  bottom: 5px;
+  right: 5px;
+  width: 25px;
+  height: 25px;
+  stroke-width: 2px;
+  stroke: #fff;
+  fill: red;
+}
