+++ src/app/ui/main/photo-list.module.css
@@ -1,0 +1,32 @@
+.photoList {
+  margin-top: 50px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  grid-template-rows: 1fr;
+  gap: 20px;
+}
+
+@media (max-width: 860px) {
+  .photoList {
+    grid-template-columns: 1fr 1fr 1fr;
+  }
+}
+
+@media (max-width: 650px) {
+  .photoList {
+    grid-template-columns: 1fr 1fr;
+  }
+}
+
+@media (max-width: 422px) {
+  .photoList {
+    grid-template-columns: 1fr;
+  }
+}
+
+.pagingWrapper {
+  border: 1px solid;
+  height: 50px;
+  margin-top: 50px;
+  margin-bottom: 100px;
+}
