+++ src/app/ui/main/search.module.css
@@ -1,0 +1,47 @@
+/* search */
+
+.searchCover {
+  height: 300px;
+  background: url("/images/cover.webp") no-repeat center;
+  background-size: cover;
+  filter: grayscale(50%);
+
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.search {
+  width: 1024px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  justify-content: center;
+  color: #fff;
+  padding: 20px;
+}
+
+.title {
+  font-size: 40px;
+  font-weight: bold;
+}
+
+.input {
+  height: 54px;
+  width: 70%;
+  border-radius: 10px;
+  border: 0;
+  padding-left: 20px;
+  color: black;
+}
+
+.searchIcon {
+  color: gray;
+  width: 23px;
+  height: 23px;
+  position: relative;
+  left: -35px;
+  top: 5px;
+  stroke-width: 2.5px;
+}
