+++ src/app/ui/modal.module.css
@@ -1,4 +1,3 @@
-/* model */
 .modalRef {
   width: 500px;
   height: 500px;
@@ -27,14 +26,11 @@
   display: flex;
   flex-direction: column;
   background-color: #fff;
-
-  /* width: calc(100% - 200px); */
   width: 80%;
-  height: 70%;
+  height: 80%;
 }
 
 .modalTitle {
-  /* border: 1px solid; */
   width: 100%;
   height: 40px;
   display: flex;
@@ -61,7 +57,6 @@
 }
 
 .modalButtons {
-  /* border: 1px solid; */
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -97,11 +92,9 @@
   margin-left: 50px;
   width: 100%;
   height: 200px;
-  /* border: 1px solid; */
 }
 
 .photoInfo {
-  /* border: 1px solid; */
   width: 100%;
   display: flex;
   flex-direction: row;
@@ -109,7 +102,6 @@
 }
 
 .infoWrapper {
-  /* border: 1px solid; */
   width: 200px;
   height: 80px;
 }
@@ -126,7 +118,6 @@
 }
 
 .tags {
-  /* border: 1px solid; */
   display: flex;
   align-items: center;
   gap: 10px;
